prevent the re-enter of fault injection exception handler
there's a few FaultInjection bug because of system exception like low
resource causing stack overflow, when fault injection infra trying to
analyze the exception, the infra itself can hit the system exception
again and re-enter the exception handler. There was a removal of the
exception handler call, however, it didn't work when the removing is
calling inside the handler. see the comments in code change for detail