Checkpoint commit: lib/common/core
Get part of lib/common/core building.
Bring in more of the PAL to support this.
Add CH_WSTR macro to support 2-byte string constants.
Disable part of CommonMin.h to get BinaryFeatureControl.cpp building- we
need to re-enable parts of it to get the rest of lib/common/core building.