toString() on ES6 class constructors returns class declaration or expression
The ES6 spec states at http://www.ecma-international.org/ecma-262/6.0/#sec-function.prototype.tostring:
If the object was defined using ECMAScript code and the returned string
representation is not in the form of a MethodDefinition or GeneratorMethod
then the representation must be such that if the string is evaluated, using
eval in a lexical context that is equivalent to the lexical context used to
create the original object, it will result in a new functionally equivalent
object.

This means ES6 class constructors, when toStringed, should return a string
in a form of either a ClassDeclaration or a ClassExpression.