Native for loop's head has destructuring declaration which must have an initializer.
Native for loop with destructuring does not enforce that it must have an initializer. For that purpose the ParseVariableDeclaration takes another param which is actually deducing whether native for loop is okay or not. We do this kind of checking as we don't know whether current loop turned out to be native loop or for..in/of loop.
If the current syntax turned out to be native for loop and during declaration we deduced that native for loop will not be okay then we throw error.
While playing with different test cases  I found that TrackAssignment assert requires adjustment as in the case of for..in loop we rescan the loop head again for validating destructuring grammar. Due that assert is not valid as the symbol span require re-adjustment.