Fix async function parameters errors in strict mode
Async functions were not observing strict mode static semantics for
their formal parameters lists, i.e. no 'arguments' nor 'eval', and no
duplicates.  The reason was that the strictness of the inner generator
(which gets the user code body) was not propagating the strict mode bit
out to the enclosing function which contains the user code
parameter list.

Propagating the bit fixes everything since we then see that the function
turned on strict mode and do the normal parameter rechecking at the end
of parsing the function.