Split Scope: This is the base changelist to enable split scopes for parameter and body scopes. Symbols in param scope and body scope needs to be treated separate. Right now we merge both scopes together. But if there is a function or eval which captures one of the formals we can no longer merge the scopes.
This changelist lays out the basic structure required for enabling these
scenarios. This only addresses the scenario where a function definition is present in
the param scope. Other scenarios will be addressed in later changelists.

When a function definition occurs in the param scope that scope is marked
to indicate that it cannot be merged with the body scope. After all the
formals are parsed one duplicate symbol is created for each formal
parameter in the body scope. When emitting the param scope it is treated
similar to a new inner scope. Emits bytecodes to copy the final value of a
formal parameter as the initial value of the corresponding symbol in the
body scope.