Fix #246 by always reporting the 'use strict' directive in the parser
There was a micro optimization in the directive parsing that didn't
bother doing a string comparison to check for the 'use strict' directive
in a function if the function was already in a strict mode context.

That used to be fine but now ES6 says that it is an error if a
function's parameter list is 'non-simple' and the function has a 'use
strict' directive.

Removing the early out and always reporting the directive fixes this bug
by letting the error detection we already had in place do its job.