Fix async function non-simple parameter lists and 'use strict' error
When we see a 'use strict' directive the ES6 spec says we must report a
syntax error if the containing function's parameter list was non-simple.

In the case of async functions, the body which contains the 'use strict'
directive' is parsed as a separate function, the inner generator
function.  This function has its own parameters that do not match the
user's parameters.  The non-simpleness of the parameter list was not
propagated down to the generator.

This change propagates it down temporarily, just for the parsing, since
it has meaning beyond the parser out in the bytecode emitter that does
not apply to this particular generator function.

To do this I changed IsSimpleParamList to HasNonSimpleParameterList and
made it a flag instead of a computed property.  This has the down side
that we must set it in all the cases that make the list non-simple, but
the benefit is we can explicitly set and unset this bit.