Bug with lossy int type spec
We dont kill type specialized values across calls for perf reasons. This
can be a problem when we lossy int type specialize an object. If a
call following the conversion overrides, for example, valueOf on the
object and the object is used as an int after the call (for example, in a
shift), we will use the already available int version and end up not
executing valueOf.

Fix is to have FromVar bailout if the source is an
object and conversion was lossy.