Fixes async function parameter redeclaration cases
The parameters of an async function were disconnected from the body in
terms of redeclaration semantics since async functions are implemented
using a parser syntax tree transformation where the body is placed
inside an inner generator function expression.  So the parameters became
related to the body as the parameters of the parent function, available
for capture, but not interacting with local bindings of the same name.

This is fixed by placing PidRefs to the parameter symbols into the
parameter scope of the inner generator function.  These refs are seen
through the normal binding mechanism and interact as though they were
local parameter symbols, but they point at the outer function's
parameter symbols so that the actual binding for bytecode generation
correctly sees them as coming from the outer function.

This required touching up the bytecode generator in the case of a local
function declaration whose name matches a parameter name due to
assumptions made that local function declaration symbols will always be
local to the enclosing function.  This is no longer true in the async
function case.