Move rarely used pointer from FunctionBody to a seperate structure. This is part of effort that reducing FunctionBody size. Currently there are a lot pointers defined on FunctionBody but rarely used. From a statistic of facebook page, the total saving is about 18% size reduction on FunctionBody. For the pointers moved in this change, a couple of them never get instantiated in my statistic, and most of them has less than 2% instantiated.
Here's the structure design:
AuxPtrsFix(16 bytes or 32 bytes) layout:
                 max count  metadata(init'd type)        pointers array
--------------------------------------------------------------------------------------------------
AuxPtr16 on x64    1 byte   1 bytes                      8 bytes to hold up to 1 pointer
AuxPtr32 on x64    1 byte   3 bytes                      24 bytes to hold up to 3 pointers
AuxPtr16 on x86    1 byte   3 bytes                      12 bytes to hold up to 3 pointers
AuxPtr32 on x64    1 byte   6 bytes                      24 bytes to hold up to 6 pointers

It's doing linear search in the getter to find the instantiated pointers.
After the max count is exceeded in above fixed structures, it will promote to a dynamic expanding structure and use table looking up search.
Here's the layout:

count      array of positions    array of instantiated pointers
----------------------------------------------------------------
1 byte     currently 21 bytes    dynamic array depending on how many pointers has been instantiated

Each time when the space is not enough, it expand by 16 bytes to a bigger bucket.