[MERGE #278] Objptr copy-prop bug
Merge pull request #278 from LouisLaf:objptr
The typescript compiler hit a nasty field copy-prop bug when dealing with objptr copy-prop.  If the symbol had been killed, we would invalidate the propertySym using the new objptr, but not the old one.  In typescript, we ended up with an invalid field copy-prop. The fix is to symply invalidate both.  No impact on octane.