Clean up and fix concurrent/partial GC macro
Renamed CONCURRENT_GC_ENABLED/PARTIAL_GC_ENABLED to ENABLE_CONCURRENT_GC/ENABLE_PARTIAL_GC
to follow the same pattern as other ENABLE_* macros in Chakra.
Converted #ifdef usage of these macros to #if so that setting them to 1 or 0 has an effect

ChakraCore can now be compiled with ENABLE_CONCURRENT_GC = 0 and ENABLE_PARTIAL_GC = 0.
Note that by default, Concurrent and Partial GCs are enabled.

Some Rescan functionality has been moved out of being scoped under Concurrent-Enabled
because Rescan is also used for marking in low-memory situations. However, the code has
been modified to not depend on write-watch behavior when Concurrent and Partial is off.
This change makes it easier to stand up the Recycler code in the linux branch.

Reviewer: @curtisman