HTML Comment Spec Compliance
Removed custom logic from HTML comment handling and brought up to date with B.1.3 HTML-like Comments.
- Removed HTML comment flag and replaced with a flag for checking if we are parsing a module (the only place we don't allow HTML comments in the standard.)
- Removed custom logic involving looking for EOF and } when parsing HTML end comment and jump to single line comment handling instead.
- Removed some scanner lookahead functions that are no longer used.
- Added negative parsing cases to modules syntax testing.

The unit test added for this commit is a special file and must be treated like binary to preserve the mixed and non-standard line endings. I added a line to .gitattributes to ensure this.

Fixes #20.