This change set implements param scope and the body scope as parallel scopes.
The previous logic of having param scope as an inner scope to the body
scope does not work well in eval cases. This changelist only addresses
function definitions in the param scope. I will be working on the eval
cases next. This works in interpreter mode right now. Backend changes are
not done yet.

Two new bytecodes were added. BeginBodyScope marks the beginning of the
body scope where a new closure object is created for the body scope. Until
then closure created for the param scope is used. After BeginBodyScope
theparam scope one is not used other than copying the initial value for
the body scope symbols from the param scope symbols. This required the
creation of another new bytecode LdParamSlot. This also include some
refactoring required for enabling these.