Lambda with default causes issue.
In order to determine that current expression is a lambda we parse till an arrow (=>) and then reparse the whole expression as lambda formal list. While reparsing we used to just reset the scanner state to an appropriate position, but no adjustment done on the parser state. Before default there was no adjustment needed for parser, but with default - we could have parsed function/classes on formals and they remain there in the AST and they have altered the parser state already and lambda parsing will just keep adding to that altered parser state.
Fixed this by storing some important variables and then restoring them back once we are about to reparse the grammar for lambda.
Writing test for this is kind of tricky as this weird state is *almost* not observable from the javascript. One of the assert fired in our internal testing and that is added as a test.
I have scoped Capturing and Restoring only for lambda scenario. In future we could merge scanner and parser state if more evidence found.