Debugger support for default parameters
When we stop at formals (params) location we need to show locals, visible
only at formals scope. For that purpose DebuggerScope is used (with a new
type DiagParamScope). That scope will contain the bytecode range and symbols. With
the help of that scope we will filter variables when we are halted at
formals location.