Update copyright check to accept a new, more inclusive, copyright.
In order to address concerns over community copyright of the source code they modify. All authors are still subject to the terms of the CLA they sign when contributing.

This new format was approved by Microsoft legal:

```
//-------------------------------------------------------------------------------------------------------
// Copyright (C) Microsoft Corporation and contributors. All rights reserved.
// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
//-------------------------------------------------------------------------------------------------------
```

In addition to the original format:

```
//-------------------------------------------------------------------------------------------------------
// Copyright (C) Microsoft. All rights reserved.
// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
//-------------------------------------------------------------------------------------------------------
```

The check now allows both so that we don't need to make a disruptive change to every file in the project. The notice can be updated at the author's discretion as they make changes.