Incorrect line and column number on error stack.
If the object or array literals are at the top level statement we didn't
enclosed them in Start/End Statement. Fixed that and fixed other places in
the Emit function wherever I can see problem.
Also for 'extends' we didn't enclose the expression in Start/End Statement so if that reproduce an error that error was reported on wrong place. Fixed that.

few more scenarios handled and added test for that.