Fixing current deferred stub issue due to default on lambda
Default as function on lambda's formal location could be problematic when
they are undeferring. When we parse those default function we don't know
that they are part of lambda so those functions will try to use
m_currDeferredStub - which was not meant for them (as it was allocated for
the lamda). I have fixed that by using the function location (ichMin) as a
checkpoint, if they match we will use the current deferred stub otherwise
it is possible that is meant for lambda.
There is no simple repro and repro involes the page heap enabled, That is
why I didn't add any unittest case for that.