Fix -0 tracking bugs in Add and Sub operations
The forward pass sets a bit, m_wasNegativeZeroPreventedByBailout, on an
operand if it was int-type-specialized and was protected by a
BailOutOnNegativeZero (which would happen if there are uses of that
operand downstream in manners which require distinction between -0 and +0
and the forward pass cannot say for sure that the operand can't be -0).
The Deadstore pass then looks at this bit to determine if it can remove
some -0 bailouts.

However, this bit being set to false for an operand doesn't necessarily mean that that
operand cannot be -0 if the operation that produced this operand wasn't type-spec'ed at all.
Negative zero tracking in the deadstore pass wasn't taking this into
account and was getting rid of more -0 bailouts than it should have,
resulting in wrong results.