Address perf regression from TypedArray constructor change related to iterable object
Recent TypedArray constructor change to support iterable object as input had unintentionally
shadowed codepaths constructing a TypedArray from an array (iterable object), causing perf regression.

Fix by adding stricter checks to keep arrays out of the new codepath designed specifically for iterable
object. This is done by retrieving `[Symbol.iterator]` property of the object and checking if it is not
the built-in Array.prototype[Symbol.iterator], or if the object is not an array but still has an iterator
(not `undefined`). Only if these conditions are met would the object be used as an iterable for constructing
the `TypedArray`. Otherwise it will go down the existing codepaths for array or error-throwing.