Fixing a subtle Switch Opt Bug
Issue:
The interpreter profiles the switch expression as a string. But the
globopt sees the value of the expression as an Integer, due to effects
of constant folding and BailOnNoProfile.

Fix:
When such incompatibility is seen, we do a compile time rejit.
This is already done when we have a sym. But we were just asserting for
this in constant folding path.