[OS5427497] Parser mistakes 'new.target' as in global function under -forceundodefer
Global function is not allowed to have 'new.target' expression. Detection of being in the global function
depends on m_currentNodeNonLamdaFunc being nullptr for the global function and pointing to an actual
function pnode otherwise. During undo of deferred parsing, however, m_currentNodeNonLambdaFunc is not
properly set to a function pnode, causing parser to mistake 'new.target' as if it is in the global function.
Fix by assigning m_currentNodeNonLambdaFunc to pnodeFnc.