Fix OS 4721262: Report fatal error when JsAddRef fails in CallbackMessage ctor.
Allowing the code to continue after a failed JsAddRef will cause corruption of the datastructure which will cause issues later, notable during the dtor when it looks for the Ref which should have been added and should exist.

When OOM occurs and this call fails, we should terminate the process because the OOM is not recoverable in this case.