Scope slot positions for destructuring patterns on param.
We were not assigning the slot position to the destructuring pattern itself. eg funciton foo({}, a), we were not assigning the slot for the pattern, so the 'a' used to get the first slot. This was creating the problem for the arguments object as while creating the arguments object we were having formals count as 2 but only 1 slot assigned so the arguments object to frameObject interaction is messed up. For that reason I have created the slots for the pattern itself.
While doing that I have also made sure the order of assigning slots should be consistent.
such as
1. normal params and patterns
2. rest
3. arguments object
4. Identifiers inside the pattern.

Since I have touched the way 'rest' used to get slots I have added tests for 'rest' as well for validation.