Class extends clause has different source locations between defer/non-defer parse modes
Tracked in OS 7030846.

When emitting a class statement with an extends clause, we made the assumption that the start and end of the bytecode for that statement happened in the same place in defer and non-defer parse modes. In fact, it is possible to have the load happen in different locations, e.g. if we cache a previous load's register location in non-defer parse mode and have to load from a local slot in defer parse mode.

The fix is to set the statement location to the InitClass bytecode.