Module script sources should be visible in the debugger
Scipts within modules were not registered to the debugger. This work enables
them to register them properly. The dwHostSourceContext was already setup
correctly, the work remaining is to create the proper sourceContextInfo
and link them to utf8sourceinfo. As we create the correct
sourceContextInfo the module code exposed the inliner bug (earlier we
were not inlining - after current work we are now) - fixed that as well.

Note : Locals experience is also not optimum for the module - but that
will be tackled in the later stages.