Fix for issue with using this object in the param scope of a lambda function
When there are function definitions in the param scope of lambda functions with split scope, those functions belong to the param scope. When we were doing the look up for a non-lambda function, like in the case of looking up for this object, we were always starting from the body which was causing issues.