Update arguments count for an inlinee at InlineeStart instead of InlineeEnd
We currently do this when we see InlineeEnd, and have special cases for
opcodes (such as InlineThrow, InlineRuntimeError) which may make
InlineeEnd unreachable. Updating the maxInlineeArgoutCount on the func
upon seeing an InlineeStart is safer and more sustainable. We wouldn't
have to keep adding cases in which InlineeEnd might get removed.